refactor: reorganize server utilities and remove legacy packages
- Remove server-only package (unnecessary with Next.js 16)
- Move server utilities to lib/utils/server/ folder
  - audit.ts
  - scheduling.ts
  - service-health.ts
  - users.ts
  - server-utils.ts
- Update all imports to use new lib/utils/server/ path
- Remove unnecessary lazy imports (trust Next.js default behavior)
- Update Jest test mocks to use new paths
- All tests passing ✅
- Build successful ✅
