+++ app/api/appointments/[meetingId]/records/route.ts
@@ -1,7 +1,7 @@
 import { db } from '@/drizzle/db';
 import { RecordTable } from '@/drizzle/schema';
-import { logAuditEvent } from '@/lib/utils/audit';
 import { decryptRecord, encryptRecord } from '@/lib/utils/encryption';
+import { logAuditEvent } from '@/lib/utils/server/audit';
 import { auth } from '@clerk/nextjs/server';
 import { eq } from 'drizzle-orm';
 import { headers } from 'next/headers';
