+++ server/actions/events.ts
@@ -3,7 +3,7 @@
 import { db } from '@/drizzle/db';
 import { EventTable, MeetingTable } from '@/drizzle/schema';
 import { getServerStripe } from '@/lib/integrations/stripe';
-import { logAuditEvent } from '@/lib/utils/audit';
+import { logAuditEvent } from '@/lib/utils/server/audit';
 import { eventFormSchema } from '@/schema/events';
 import { checkExpertSetupStatus, markStepComplete } from '@/server/actions/expert-setup';
 import { EVENT_CREATED, EVENT_DELETED, EVENT_UPDATED } from '@/types/audit';
